#13 RegisterCommentRequest를 vo패키지로
