+++ src/main/java/site/moasis/monolithicbe/application/controller/CommentController.java
@@ -2,22 +2,16 @@ package site.moasis.monolithicbe.application.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import site.moasis.monolithicbe.common.response.CommonResponse;
-import site.moasis.monolithicbe.domain.comment.service.CommentCommand;
-import site.moasis.monolithicbe.domain.comment.service.CommentInfo;
 import site.moasis.monolithicbe.domain.comment.service.CommentReadService;
 import site.moasis.monolithicbe.domain.comment.service.CommentWriteService;
-
 import java.util.UUID;
-
-import static site.moasis.monolithicbe.domain.comment.dto.CommentDto.CommentCreateDto;
+import static site.moasis.monolithicbe.domain.comment.vo.CommentVo.*;
+import static site.moasis.monolithicbe.domain.comment.service.CommentCommand.*;
 
 @RestController
 @RequiredArgsConstructor
@@ -28,36 +22,36 @@ public class CommentController {
     private final CommentReadService readService;
     private final CommentWriteService writeService;
 
-    @GetMapping
+    @GetMapping("/comments")
     @Operation(summary = "댓글 전체 조회")
-    public ResponseEntity<CommonResponse<?>> findAll(){
+    public ResponseEntity<?> findAll() {
         var commentInfo = readService.selectAll();
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(CommonResponse.success(commentInfo, "전체 댓글 조회 성공"));
     }
 
-    @GetMapping("/{commentId}")
+    @GetMapping("/comment/{commentId}")
     @Operation(summary = "댓글 단건 조회")
-    public ResponseEntity<CommonResponse<?>> findOne(@PathVariable("commentId") UUID commentId){
+    public ResponseEntity<?> findOne(@PathVariable("commentId") UUID commentId) {
         var commentInfo = readService.selectOne(commentId);
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(CommonResponse.success(commentInfo, "댓글 조회 성공"));
     }
 
-    @GetMapping("/users/{userId}")
+    @GetMapping("/comments/users/{userId}")
     @Operation(summary = "user_id를 통한 댓글 조회")
-    public ResponseEntity<CommonResponse<?>> findByUser(@PathVariable Long userId){
+    public ResponseEntity<?> findByUser(@PathVariable UUID userId) {
         var commentInfo = readService.selectByUser(userId);
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(CommonResponse.success(commentInfo, "Id가 " + userId + "인 유저가 작성한 댓글 조회 성공"));
     }
 
-    @GetMapping("/articles/{articleId}")
+    @GetMapping("/comments/articles/{articleId}")
     @Operation(summary = "article_id를 통한 댓글 조회")
-    public ResponseEntity<CommonResponse<?>> findByArticle(@PathVariable Long articleId){
+    public ResponseEntity<?> findByArticle(@PathVariable UUID articleId) {
         var commentInfo = readService.selectByArticle(articleId);
         return ResponseEntity
                 .status(HttpStatus.OK)
@@ -66,16 +60,14 @@ public class CommentController {
 
     @PostMapping("/users/{userId}/articles/{articleId}/comments")
     @Operation(summary = "댓글 생성")
-    public ResponseEntity<CommonResponse<CommentInfo>> registerComment(
+    public ResponseEntity<?> registerComment(
             @PathVariable UUID userId,
             @PathVariable UUID articleId,
-            @RequestBody RegisterCommentRequest registerCommentRequest){
+            @RequestBody RegisterCommentRequest registerCommentRequest) {
 
-        UUID loginUserId = userId;
-
-        var command = CommentCommand.RegisterCommentCommand
+        var command = RegisterCommentCommand
                 .builder()
-                .userId(loginUserId)
+                .userId(userId)
                 .articleId(articleId)
                 .content(registerCommentRequest.getContent())
                 .build();
@@ -87,37 +79,12 @@ public class CommentController {
                 .body(CommonResponse.success(commentInfo, "댓글 생성 완료"));
     }
 
-    @Getter
-    @AllArgsConstructor
-    @NoArgsConstructor
-    public static class RegisterCommentRequest {
-        private String content;
-    }
-
-    @DeleteMapping("/id/{commentId}")
+    @DeleteMapping("/comment/id/{commentId}")
     @Operation(summary = "id를 통한 댓글 삭제")
-    public ResponseEntity<CommonResponse<?>> deleteOne(@PathVariable("commentId") UUID commentId){
+    public ResponseEntity<?> deleteOne(@PathVariable("commentId") UUID commentId) {
         var commentInfo = writeService.dropOne(commentId);
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(CommonResponse.success(null, commentInfo + "번 댓글 삭제 완료"));
     }
-
-//    @PostMapping("/user/{userId}")
-//    @Operation(summary = "해당 user가 작성한 댓글 삭제")
-//    public ResponseEntity<CommonResponse<?>> deleteByUser(@PathVariable Long userId){
-//        var commentInfo = writeService.dropByUser(userId);
-//        return ResponseEntity
-//                .status(HttpStatus.OK)
-//                .body(CommonResponse.success(null, "Id가 " + commentInfo + "인 유저가 작성한 댓글 삭제 성공"));
-//    }
-
-//    @PostMapping("/article/{articleId}")
-//    @Operation(summary = "해당 article에 속한 댓글 삭제")
-//    public ResponseEntity<CommonResponse<?>> deleteByArticle(@PathVariable Long articleId){
-//        var articleInfo = writeService.dropByArticle(articleId);
-//        return ResponseEntity
-//                .status(HttpStatus.OK)
-//                .body(CommonResponse.success(null, articleInfo + "번 게시물의 댓글 삭제 성공"));
-//    }
 }
